move &ibalancer.ConnectivityStateEvaluator
